+++ src/main/java/pl/equipment/store/domain/product/Product.java
@@ -4,18 +4,20 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor(access = AccessLevel.PACKAGE)
 @Getter
 class Product {
     private final Long id;
     private final String name;
     private final String description;
-    private final double unitPrice;
+    private final BigDecimal unitPrice;
     private final int unitsInStock;
     private final String imageUrl;
     private final boolean active;
 
-    static Product create(String productName, String description, double unitPrice, int unitsInStock){
+    static Product create(String productName, String description, BigDecimal unitPrice, int unitsInStock){
         return new Product(null, productName, description, unitPrice, unitsInStock, "", true);
     }
