+++ src/main/java/pl/equipment/store/domain/product/dto/CreateProductDto.java
@@ -3,11 +3,13 @@ package pl.equipment.store.domain.product.dto;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigDecimal;
+
 @RequiredArgsConstructor
 @Getter
 public class CreateProductDto {
     private final String name;
     private final String description;
-    private final double unitPrice;
+    private final BigDecimal unitPrice;
     private final int unitsInStock;
 }
